Store offer price as integer grosze instead of decimal
Offer.price was a decimal column mapped to a JS number, which risks
floating-point rounding errors on money. Switched the column to
integer (smallest currency unit) with a migration that converts
existing values (price * 100, rounded). The admin form still shows
and accepts złoty - it converts to/from grosze at the form boundary -
and every place that renders a price divides by 100 before formatting.

The migration hasn't been run against any real database yet; back up
before applying it on production.

+++ api/src/modules/offer/offer.entity.ts
@@ -59,7 +59,8 @@ export class Offer {
   @Column({ type: 'uuid', nullable: false })
   shipId: string;
 
-  @Column('decimal')
+  // Cena w groszach (najmniejsza jednostka waluty) - nie w złotych.
+  @Column('integer')
   price: number;
 
   @Column()
